+++ climpred/alignment.py
@@ -1,3 +1,5 @@
+from typing import List, Optional, Union
+
 import dask
 import numpy as np
 import xarray as xr
@@ -8,7 +10,13 @@ from .exceptions import CoordinateError
 from .utils import get_multiple_lead_cftime_shift_args, shift_cftime_index
 
 
-def return_inits_and_verif_dates(forecast, verif, alignment, reference=None, hist=None):
+def return_inits_and_verif_dates(
+    forecast: xr.Dataset,
+    verif: xr.Dataset,
+    alignment: Union[str, List[str]],
+    reference=Union[List[str], str],
+    hist=Optional[xr.Dataset],
+):
     """Returns initializations and verification dates for an arbitrary number of leads
     per a given alignment strategy.